+++ src/components/shared/Carousel/index.tsx
@@ -8,13 +8,9 @@ interface CarouselProps {
 export default function Carousel({ images }: CarouselProps) {
     const [currentIndex, setCurrentIndex] = useState(0);
 
-    const goToPrevious = () => {
-        setCurrentIndex((prevIndex) => prevIndex === 0 ? images.length - 1 : prevIndex - 1);
-    };
+    const goToPrevious = () => setCurrentIndex((prevIndex) => prevIndex === 0 ? images.length - 1 : prevIndex - 1);
 
-    const goToNext = () => {
-        setCurrentIndex((prevIndex) => prevIndex === images.length - 1 ? 0 : prevIndex + 1);
-    };
+    const goToNext = () => setCurrentIndex((prevIndex) => prevIndex === images.length - 1 ? 0 : prevIndex + 1);;
 
     return (
         <CarouselContainer>