+++ src/data/projects.ts
@@ -10,16 +10,14 @@ function getProjectImages(dir: string, numberOfImages: number): string[] {
 
 const projects = [
     {
-        title: 'Igreja Vivadoração',
-        type: 'website',
-        technologies: ['wordpress', 'php', 'html', 'css', 'javascript', 'mysql'],
-        cardImage: 'images/projects/vivadoracao/0.png',
-        description: `Site institucional da "Igreja Vivadoração". Possui informações sobre a estrutura da igreja, 
-        além de informações sobre os eventos, ministérios e redes da igreja. Também inclui informações sobre a 
-        membresia da igreja, além de formulários de inscrição para eventos, membresia e voluntariado. Site 
-        responsivo.`,
-        images: getProjectImages('vivadoracao', 17),
-        liveProject: 'https://vivadoracao.com.br'
+        title: 'MIDI Viewer',
+        type: 'webapplication',
+        technologies: ['react', 'html', 'css', 'javascript'],
+        cardImage: 'images/projects/midi-viewer/3.png',
+        description: `Um visualizador de arquivos MIDI em um piano virtual. Funcionalidade similar ao "Synthesia".`,
+        images: getProjectImages('midi-viewer', 7),
+        liveProject: 'https://natanaelnogueira144.github.io/midi-viewer',
+        repositoryUrl: 'https://github.com/NatanaelNogueira144/midi-viewer'
     },
     {
         title: 'Carteira Digital',
@@ -32,25 +30,6 @@ const projects = [
         repositoryUrl: 'https://github.com/NatanaelNogueira144/carteira-digital-frontend'
     },
     {
-        title: 'Espaço Lírio Visitas',
-        type: 'webapplication',
-        technologies: ['php', 'html', 'css', 'javascript', 'jquery', 'bootstrap', 'mysql'],
-        cardImage: 'images/projects/espaco-lirio-visitas/1.png',
-        description: `Aplicação que tem a finalidade de gerenciar visitas, eventos e gerar contratos automatizados 
-        do "Espaço Lírio Eventos". Inclui autenticação, sistemas CRUD e responsividade, além de assinatura digital.`,
-        images: getProjectImages('espaco-lirio-visitas', 10)
-    },
-    {
-        title: 'Smart House Aluguéis',
-        type: 'webapplication',
-        technologies: ['php', 'html', 'css', 'javascript', 'jquery', 'bootstrap', 'mysql'],
-        cardImage: 'images/projects/smart-houses-alugueis/1.png',
-        description: `Aplicação que tem a finalidade de gerenciar aluguéis, faturas e pagamentos, além de gerar 
-        contratos automatizados das casas de aluguel do "Smart Houses". Inclui autenticação, sistemas CRUD e 
-        responsividade, além de assinatura digital.`,
-        images: getProjectImages('smart-houses-alugueis', 6)
-    },
-    {
         title: 'API da Bíblia',
         type: 'api',
         technologies: ['laravel', 'php', 'html', 'css', 'mysql'],
@@ -59,36 +38,6 @@ const projects = [
         apresentar dois planos diferentes de leitura bíblica anual. É de fácil integração com outros sistemas.`,
         images: getProjectImages('bible-api', 8),
         repositoryUrl: 'https://github.com/NatanaelNogueira144/bible-api'
-    },
-    {
-        title: 'IPPV',
-        type: 'website',
-        technologies: ['wordpress', 'php', 'html', 'css', 'javascript', 'mysql'],
-        cardImage: 'images/projects/ippv/0.png',
-        description: `Site institucional da empresa de treinamentos "Instituto Princípios Para Viver". 
-        Inclui informações da empresa, como seus serviços, trainers e eventos, além de responsividade.`,
-        images: getProjectImages('ippv', 5),
-        liveProject: 'https://ippv.com.br'
-    },
-    {
-        title: 'IPPV Coach',
-        type: 'website',
-        technologies: ['wordpress', 'php', 'html', 'css', 'javascript', 'mysql'],
-        cardImage: 'images/projects/ippv-coach/0.png',
-        description: `Site de coaching da empresa "Instituto Princípios Para Viver". Inclui informações dos 
-        treinamentos, trainers e eventos, além de um blog com posts da Harvard Business Review. Site responsivo.`,
-        images: getProjectImages('ippv-coach', 6),
-        liveProject: 'https://ippv.com.br/coach'
-    },
-    {
-        title: 'IPPV Master',
-        type: 'website',
-        technologies: ['wordpress', 'php', 'html', 'css', 'javascript', 'mysql'],
-        cardImage: 'images/projects/ippv-master/0.png',
-        description: `Site da masterclass da empresa "Instituto Princípios Para Viver". Inclui informações sobre a 
-        masterclass, trainers e localização. Site responsivo.`,
-        images: getProjectImages('ippv-master', 3),
-        liveProject: 'https://ippv.com.br/master'
     },
     {
         title: 'Teste das Cinco Linguagens do Amor',
@@ -109,16 +58,6 @@ const projects = [
         Inclui autenticação, sistemas CRUD e responsividade.`,
         images: getProjectImages('teste-inteligencia-espiritual', 3),
         repositoryUrl: 'https://github.com/NatanaelNogueira144/teste-inteligencia-espiritual'
-    },
-    {
-        title: 'MIDI Viewer',
-        type: 'webapplication',
-        technologies: ['react', 'html', 'css', 'javascript'],
-        cardImage: 'images/projects/midi-viewer/3.png',
-        description: `Um visualizador de arquivos MIDI em um piano virtual. Funcionalidade similar ao "Synthesia".`,
-        images: getProjectImages('midi-viewer', 7),
-        liveProject: 'https://natanaelnogueira144.github.io/midi-viewer',
-        repositoryUrl: 'https://github.com/NatanaelNogueira144/midi-viewer'
     },
     {
         title: 'KojiSplit',